Make conditional routing based on localStorage
I did this as without persistence the user will be automatically logged out on refresh. I did this by reintroducing the userLoggedIn state variable and making the routes conditionally render based on this variable instead of the username state variable. Now when the user successfully logs in, the localStorage value gets set to true