Change setUserLoggedIn behaviour
Change everywhere in the code that sets the localStorage value 'userLoggedIn' by moving the code into a useEffect hook. This is possible as mostly everywhere the 'userLoggedIn' value is set, the state variable username is also changed, so username is used as a dependency to the useEffect. I did this as it is good practice to handle utilities outside of React inside of a useEffect hook for better performance, especially on slower devices