Replace function that saves user to localStorage
Replace saveUser function with useEffect hook by extended an already existing useEffect that handles saving setUserLoggedIn to localStorage. I did this by checking if the username is the same as any keys in the localStorage. If it is not then it is a new user, and then we need to save the current input field values for the username and password on the signup page. I put this in a useEffect as it improves user experience through performance being delayed to after the visual React render, especially on weak networks and reduces code